@@ -11,6 +11,10 @@
         {
             "name": "Adam Wathan",
             "email": "adam.wathan@gmail.com"
+        },
+        {
+            "name": "Matt Stauffer",
+            "email": "matt@tighten.co"
         }
     ],
     "autoload": {
@@ -23,24 +27,34 @@
             "Valet\\": "cli/Valet/"
         }
     },
+    "autoload-dev": {
+        "files": [
+            "tests/UsesNullWriter.php",
+            "tests/BaseApplicationTestCase.php",
+            "tests/Drivers/BaseDriverTestCase.php"
+        ]
+    },
     "require": {
-        "php": "^7.0|^8.0",
-        "illuminate/container": "~5.1|^6.0|^7.0|^8.0|^9.0",
+        "php": "^7.1|^8.0",
+        "illuminate/collections": "^8.0|^9.0|^10.0",
+        "illuminate/container": "~5.1|^6.0|^7.0|^8.0|^9.0|^10.0",
         "mnapoli/silly": "^1.0",
+        "symfony/console": "^3.0|^4.0|^5.0|^6.0",
         "symfony/process": "^3.0|^4.0|^5.0|^6.0",
-        "tightenco/collect": "^5.3|^6.0|^7.0|^8.0",
-        "guzzlehttp/guzzle": "^6.0|^7.4"
+        "guzzlehttp/guzzle": "^6.0|^7.4",
+        "symfony/event-dispatcher": "^3.0|^4.0|^5.0|^6.0"
     },
     "require-dev": {
         "mockery/mockery": "^1.2.3",
-        "yoast/phpunit-polyfills": "^0.2.0"
+        "yoast/phpunit-polyfills": "^0.2.0",
+        "laravel/pint": "^1.4"
     },
     "bin": [
         "valet"
     ],
     "extra": {
         "branch-alias": {
-            "dev-master": "2.x-dev"
+            "dev-master": "4.x-dev"
         }
     }
 }
