@@ -7,70 +7,49 @@ use Symfony\Component\Process\Process;
 class CommandLine
 {
     /**
-     * Simple global function to run commands.
-     *
-     * @param  string  $command
-     * @return void
+     * Simple global function to run commands quietly.
      */
-    public function quietly($command)
+    public function quietly(string $command): void
     {
         $this->runCommand($command.' > /dev/null 2>&1');
     }
 
     /**
      * Simple global function to run commands.
-     *
-     * @param  string  $command
-     * @return void
      */
-    public function quietlyAsUser($command)
+    public function quietlyAsUser(string $command): void
     {
         $this->quietly('sudo -u "'.user().'" '.$command.' > /dev/null 2>&1');
     }
 
     /**
      * Pass the command to the command line and display the output.
-     *
-     * @param  string  $command
-     * @return void
      */
-    public function passthru($command)
+    public function passthru(string $command): void
     {
         passthru($command);
     }
 
     /**
      * Run the given command as the non-root user.
-     *
-     * @param  string  $command
-     * @param  callable  $onError
-     * @return string
      */
-    public function run($command, callable $onError = null)
+    public function run(string $command, callable $onError = null): string
     {
         return $this->runCommand($command, $onError);
     }
 
     /**
      * Run the given command.
-     *
-     * @param  string  $command
-     * @param  callable  $onError
-     * @return string
      */
-    public function runAsUser($command, callable $onError = null)
+    public function runAsUser(string $command, callable $onError = null): string
     {
         return $this->runCommand('sudo -u "'.user().'" '.$command, $onError);
     }
 
     /**
      * Run the given command.
-     *
-     * @param  string  $command
-     * @param  callable  $onError
-     * @return string
      */
-    public function runCommand($command, callable $onError = null)
+    public function runCommand(string $command, callable $onError = null): string
     {
         $onError = $onError ?: function () {
         };
