@@ -3,48 +3,23 @@
 namespace Valet;
 
 use DomainException;
+use Illuminate\Support\Collection;
 
 class PhpFpm
 {
-    public $brew;
-    public $cli;
-    public $files;
-    public $config;
-    public $site;
-    public $nginx;
-
     public $taps = [
         'homebrew/homebrew-core',
         'shivammathur/php',
     ];
 
-    /**
-     * Create a new PHP FPM class instance.
-     *
-     * @param  Brew  $brew
-     * @param  CommandLine  $cli
-     * @param  Filesystem  $files
-     * @param  Configuration  $config
-     * @param  Site  $site
-     * @param  Nginx  $nginx
-     * @return void
-     */
-    public function __construct(Brew $brew, CommandLine $cli, Filesystem $files, Configuration $config, Site $site, Nginx $nginx)
+    public function __construct(public Brew $brew, public CommandLine $cli, public Filesystem $files, public Configuration $config, public Site $site, public Nginx $nginx)
     {
-        $this->cli = $cli;
-        $this->brew = $brew;
-        $this->files = $files;
-        $this->config = $config;
-        $this->site = $site;
-        $this->nginx = $nginx;
     }
 
     /**
      * Install and configure PhpFpm.
-     *
-     * @return void
      */
-    public function install()
+    public function install(): void
     {
         info('Installing and configuring phpfpm...');
 
@@ -68,10 +43,8 @@ class PhpFpm
 
     /**
      * Forcefully uninstall all of Valet's supported PHP versions and configurations.
-     *
-     * @return void
      */
-    public function uninstall()
+    public function uninstall(): void
     {
         $this->brew->uninstallAllPhpVersions();
         rename(BREW_PREFIX.'/etc/php', BREW_PREFIX.'/etc/php-valet-bak'.time());
@@ -82,11 +55,8 @@ class PhpFpm
      * Create (or re-create) the PHP FPM configuration files.
      *
      * Writes FPM config file, pointing to the correct .sock file, and log and ini files.
-     *
-     * @param  string  $phpVersion
-     * @return void
      */
-    public function createConfigurationFiles($phpVersion)
+    public function createConfigurationFiles(string $phpVersion): void
     {
         info("Updating PHP configuration for {$phpVersion}...");
 
@@ -104,7 +74,7 @@ class PhpFpm
         $contents = str_replace(
             ['VALET_USER', 'VALET_HOME_PATH', 'valet.sock'],
             [user(), VALET_HOME_PATH, self::fpmSockName($phpVersion)],
-            $this->files->get(__DIR__.'/../stubs/etc-phpfpm-valet.conf')
+            $this->files->getStub('etc-phpfpm-valet.conf')
         );
         $this->files->put($fpmConfigFile, $contents);
 
@@ -114,13 +84,13 @@ class PhpFpm
 
         $this->files->putAsUser(
             $destDir.'/php-memory-limits.ini',
-            $this->files->get(__DIR__.'/../stubs/php-memory-limits.ini')
+            $this->files->getStub('php-memory-limits.ini')
         );
 
         $contents = str_replace(
             ['VALET_USER', 'VALET_HOME_PATH'],
             [user(), VALET_HOME_PATH],
-            $this->files->get(__DIR__.'/../stubs/etc-phpfpm-error_log.ini')
+            $this->files->getStub('etc-phpfpm-error_log.ini')
         );
         $this->files->putAsUser($destDir.'/error_log.ini', $contents);
 
@@ -131,21 +101,16 @@ class PhpFpm
 
     /**
      * Restart the PHP FPM process (if one specified) or processes (if none specified).
-     *
-     * @param  string|null  $phpVersion
-     * @return void
      */
-    public function restart($phpVersion = null)
+    public function restart(?string $phpVersion = null): void
     {
         $this->brew->restartService($phpVersion ?: $this->utilizedPhpVersions());
     }
 
     /**
      * Stop the PHP FPM process.
-     *
-     * @return void
      */
-    public function stop()
+    public function stop(): void
     {
         call_user_func_array(
             [$this->brew, 'stopService'],
@@ -155,11 +120,8 @@ class PhpFpm
 
     /**
      * Get the path to the FPM configuration file for the current PHP version.
-     *
-     * @param  string|null  $phpVersion
-     * @return string
      */
-    public function fpmConfigPath($phpVersion = null)
+    public function fpmConfigPath(?string $phpVersion = null): string
     {
         if (! $phpVersion) {
             $phpVersion = $this->brew->linkedPhp();
@@ -174,7 +136,7 @@ class PhpFpm
     /**
      * Stop only the running php services.
      */
-    public function stopRunning()
+    public function stopRunning(): void
     {
         $this->brew->stopService(
             $this->brew->getAllRunningServices()
@@ -187,11 +149,8 @@ class PhpFpm
 
     /**
      * Stop a given PHP version, if that specific version isn't being used globally or by any sites.
-     *
-     * @param  string|null  $phpVersion
-     * @return void
      */
-    public function stopIfUnused($phpVersion = null)
+    public function stopIfUnused(?string $phpVersion = null): void
     {
         if (! $phpVersion) {
             return;
@@ -206,12 +165,8 @@ class PhpFpm
 
     /**
      * Isolate a given directory to use a specific version of PHP.
-     *
-     * @param  string  $directory
-     * @param  string  $version
-     * @return void
      */
-    public function isolateDirectory($directory, $version)
+    public function isolateDirectory(string $directory, string $version): void
     {
         $site = $this->site->getSiteUrl($directory);
 
@@ -233,11 +188,8 @@ class PhpFpm
 
     /**
      * Remove PHP version isolation for a given directory.
-     *
-     * @param  string  $directory
-     * @return void
      */
-    public function unIsolateDirectory($directory)
+    public function unIsolateDirectory(string $directory): void
     {
         $site = $this->site->getSiteUrl($directory);
 
@@ -252,10 +204,8 @@ class PhpFpm
 
     /**
      * List all directories with PHP isolation configured.
-     *
-     * @return \Illuminate\Support\Collection
      */
-    public function isolatedDirectories()
+    public function isolatedDirectories(): Collection
     {
         return $this->nginx->configuredSites()->filter(function ($item) {
             return strpos($this->files->get(VALET_HOME_PATH.'/Nginx/'.$item), ISOLATED_PHP_VERSION) !== false;
@@ -266,12 +216,8 @@ class PhpFpm
 
     /**
      * Use a specific version of PHP globally.
-     *
-     * @param  string  $version
-     * @param  bool  $force
-     * @return string|void
      */
-    public function useVersion($version, $force = false)
+    public function useVersion(string $version, bool $force = false): ?string
     {
         $version = $this->validateRequestedVersion($version);
 
@@ -312,11 +258,8 @@ class PhpFpm
 
     /**
      * Symlink (Capistrano-style) a given Valet.sock file to be the primary valet.sock.
-     *
-     * @param  string  $phpVersion
-     * @return void
      */
-    public function symlinkPrimaryValetSock($phpVersion)
+    public function symlinkPrimaryValetSock(string $phpVersion): void
     {
         $this->files->symlinkAsUser(VALET_HOME_PATH.'/'.$this->fpmSockName($phpVersion), VALET_HOME_PATH.'/valet.sock');
     }
@@ -324,18 +267,15 @@ class PhpFpm
     /**
      * If passed php7.4, or php74, 7.4, or 74 formats, normalize to php@7.4 format.
      */
-    public function normalizePhpVersion($version)
+    public function normalizePhpVersion(?string $version): string
     {
         return preg_replace('/(?:php@?)?([0-9+])(?:.)?([0-9+])/i', 'php@$1.$2', (string) $version);
     }
 
     /**
      * Validate the requested version to be sure we can support it.
-     *
-     * @param  string  $version
-     * @return string
      */
-    public function validateRequestedVersion($version)
+    public function validateRequestedVersion(string $version): string
     {
         if (is_null($version)) {
             throw new DomainException("Please specify a PHP version (try something like 'php@8.1')");
@@ -362,11 +302,8 @@ class PhpFpm
 
     /**
      * Get FPM sock file name for a given PHP version.
-     *
-     * @param  string|null  $phpVersion
-     * @return string
      */
-    public static function fpmSockName($phpVersion = null)
+    public static function fpmSockName(?string $phpVersion = null): string
     {
         $versionInteger = preg_replace('~[^\d]~', '', $phpVersion);
 
@@ -376,10 +313,8 @@ class PhpFpm
     /**
      * Get a list including the global PHP version and allPHP versions currently serving "isolated sites" (sites with
      * custom Nginx configs pointing them to a specific PHP version).
-     *
-     * @return array
      */
-    public function utilizedPhpVersions()
+    public function utilizedPhpVersions(): array
     {
         $fpmSockFiles = $this->brew->supportedPhpVersions()->map(function ($version) {
             return self::fpmSockName($this->normalizePhpVersion($version));
