@@ -1,7 +1,7 @@
 <?php
 
+// Allow bypassing these checks if using Valet in a non-CLI app
 if (php_sapi_name() !== 'cli') {
-    // Allow bypassing these checks if using Valet in a non-CLI app
     return;
 }
 
@@ -16,8 +16,8 @@ if (PHP_OS !== 'Darwin' && ! $inTestingEnvironment) {
     exit(1);
 }
 
-if (version_compare(PHP_VERSION, '7.0', '<')) {
-    echo 'Valet requires PHP 7.0 or later.';
+if (version_compare(PHP_VERSION, '8.0', '<')) {
+    echo 'Valet requires PHP 8.0 or later.';
 
     exit(1);
 }
