@@ -6,22 +6,16 @@ class Facade
 {
     /**
      * The key for the binding in the container.
-     *
-     * @return string
      */
-    public static function containerKey()
+    public static function containerKey(): string
     {
         return 'Valet\\'.basename(str_replace('\\', '/', get_called_class()));
     }
 
     /**
      * Call a non-static method on the facade.
-     *
-     * @param  string  $method
-     * @param  array  $parameters
-     * @return mixed
      */
-    public static function __callStatic($method, $parameters)
+    public static function __callStatic(string $method, array $parameters): mixed
     {
         $resolvedInstance = Container::getInstance()->make(static::containerKey());
 
@@ -38,6 +32,9 @@ class Nginx extends Facade
 class CommandLine extends Facade
 {
 }
+class Composer extends Facade
+{
+}
 class Configuration extends Facade
 {
 }
@@ -45,6 +42,9 @@ class Diagnose extends Facade
 {
 }
 class DnsMasq extends Facade
+{
+}
+class Expose extends Facade
 {
 }
 class Filesystem extends Facade
@@ -57,6 +57,12 @@ class PhpFpm extends Facade
 {
 }
 class Site extends Facade
+{
+}
+class Status extends Facade
+{
+}
+class Upgrader extends Facade
 {
 }
 class Valet extends Facade
