@@ -12,11 +12,14 @@ use function Valet\user;
 
 class SiteTest extends Yoast\PHPUnitPolyfills\TestCases\TestCase
 {
+    use UsesNullWriter;
+
     public function set_up()
     {
         $_SERVER['SUDO_USER'] = user();
 
         Container::setInstance(new Container);
+        $this->setNullWriter();
     }
 
     public function tear_down()
@@ -596,6 +599,12 @@ class SiteTest extends Yoast\PHPUnitPolyfills\TestCases\TestCase
                 'url' => 'http://site2.test',
                 'path' => '/Users/name/code/site2',
             ],
+            'portal.test-site' => [
+                'site' => 'portal.test-site',
+                'secured' => 'X',
+                'url' => 'http://portal.test-site.test',
+                'path' => '/Users/name/code/portal.test-site',
+            ],
         ]));
 
         $siteMock->shouldReceive('host')->andReturn('site1');
@@ -610,6 +619,9 @@ class SiteTest extends Yoast\PHPUnitPolyfills\TestCases\TestCase
 
         $this->assertEquals('site2.test', $site->getSiteUrl('site2'));
         $this->assertEquals('site2.test', $site->getSiteUrl('site2.test'));
+
+        $this->assertEquals('portal.test-site.test', $site->getSiteUrl('portal.test-site'));
+        $this->assertEquals('portal.test-site.test', $site->getSiteUrl('portal.test-site.test'));
     }
 
     public function test_it_throws_getting_nonexistent_site()
@@ -761,9 +773,9 @@ class SiteTest extends Yoast\PHPUnitPolyfills\TestCases\TestCase
 
         // When no Nginx file exists, it will create a new config file from the template
         $files->shouldReceive('exists')->once()->with($siteMock->nginxPath('site2.test'))->andReturn(false);
-        $files->shouldReceive('get')
+        $files->shouldReceive('getStub')
             ->once()
-            ->with(dirname(__DIR__).'/cli/Valet/../stubs/site.valet.conf')
+            ->with('site.valet.conf')
             ->andReturn(file_get_contents(__DIR__.'/../cli/stubs/site.valet.conf'));
 
         $files->shouldReceive('putAsUser')
@@ -888,77 +900,86 @@ class SiteTest extends Yoast\PHPUnitPolyfills\TestCases\TestCase
         $this->assertSame(['helloworld.tld'], $sites);
     }
 
+    public function test_it_returns_true_if_a_site_is_secured()
+    {
+        $files = Mockery::mock(Filesystem::class);
+        $files->shouldReceive('scandir')
+            ->once()
+            ->andReturn(['helloworld.tld.crt', '.DS_Store']);
+
+        $config = Mockery::mock(Configuration::class);
+        $config->shouldReceive('read')
+            ->once()
+            ->andReturn(['tld' => 'tld']);
+
+        swap(Filesystem::class, $files);
+        swap(Configuration::class, $config);
+
+        $site = resolve(Site::class);
+
+        $this->assertTrue($site->isSecured('helloworld'));
+    }
+
+    public function test_it_can_read_valet_rc_files()
+    {
+        resolve(Configuration::class)->addPath(__DIR__.'/fixtures/Parked/Sites');
+        $site = resolve(Site::class);
+
+        $this->assertEquals([
+            'item' => 'value',
+            'php' => 'php@8.0',
+            'other_item' => 'othervalue',
+        ], $site->valetRc('site-w-valetrc-1'));
+
+        $this->assertEquals([
+            'php' => 'php@8.1',
+        ], $site->valetRc('site-w-valetrc-2'));
+
+        $this->assertEquals([
+            'item' => 'value',
+            'php' => 'php@8.2',
+        ], $site->valetRc('site-w-valetrc-3'));
+    }
+
     public function test_it_can_read_php_rc_version()
     {
-        $config = Mockery::mock(Configuration::class);
-        $files = Mockery::mock(Filesystem::class);
+        resolve(Configuration::class)->addPath(__DIR__.'/fixtures/Parked/Sites');
+        $site = resolve(Site::class);
 
-        swap(Configuration::class, $config);
-        swap(Filesystem::class, $files);
-
-        $siteMock = Mockery::mock(Site::class, [
-            resolve(Brew::class),
-            resolve(Configuration::class),
-            resolve(CommandLine::class),
-            resolve(Filesystem::class),
-        ])->makePartial();
-
-        swap(Site::class, $siteMock);
-
-        $config->shouldReceive('read')
-            ->andReturn(['tld' => 'test', 'loopback' => VALET_LOOPBACK, 'paths' => []]);
-
-        $siteMock->shouldReceive('parked')
-            ->andReturn(collect([
-                'site1' => [
-                    'site' => 'site1',
-                    'secured' => '',
-                    'url' => 'http://site1.test',
-                    'path' => '/Users/name/code/site1',
-                ],
-            ]));
-
-        $siteMock->shouldReceive('links')->andReturn(collect([
-            'site2' => [
-                'site' => 'site2',
-                'secured' => 'X',
-                'url' => 'http://site2.test',
-                'path' => '/Users/name/some-other-directory/site2',
-            ],
-        ]));
-
-        $files->shouldReceive('exists')->with('/Users/name/code/site1/.valetphprc')->andReturn(true);
-        $files->shouldReceive('get')->with('/Users/name/code/site1/.valetphprc')->andReturn('php@8.1');
-
-        $files->shouldReceive('exists')->with('/Users/name/some-other-directory/site2/.valetphprc')->andReturn(true);
-        $files->shouldReceive('get')->with('/Users/name/some-other-directory/site2/.valetphprc')->andReturn('php@8.0');
-
-        $this->assertEquals('php@8.1', $siteMock->phpRcVersion('site1'));
-        $this->assertEquals('php@8.0', $siteMock->phpRcVersion('site2'));
-        $this->assertEquals(null, $siteMock->phpRcVersion('site3')); // Site doesn't exists
+        $this->assertEquals('php@8.1', $site->phpRcVersion('site-w-valetphprc-1'));
+        $this->assertEquals('php@8.0', $site->phpRcVersion('site-w-valetphprc-2'));
+        $this->assertEquals(null, $site->phpRcVersion('my-best-site'));
+        $this->assertEquals(null, $site->phpRcVersion('non-existent-site'));
+        $this->assertEquals('php@8.0', $site->phpRcVersion('site-w-valetrc-1'));
+        $this->assertEquals('php@8.1', $site->phpRcVersion('site-w-valetrc-2'));
+        $this->assertEquals('php@8.2', $site->phpRcVersion('site-w-valetrc-3'));
+        $this->assertEquals('php@8.2', $site->phpRcVersion('blabla', __DIR__.'/fixtures/Parked/Sites/site-w-valetrc-3'));
     }
 }
 
 class CommandLineFake extends CommandLine
 {
-    public function runCommand($command, callable $onError = null)
+    public function runCommand(string $command, callable $onError = null): string
     {
         // noop
         //
-        // This let's us pretend like every command executes correctly
+        // This lets us pretend like every command executes correctly
         // so we can (elsewhere) ensure the things we meant to do
         // (like "create a certificate") look like they
         // happened without actually running any
         // commands for real.
+
+        return 'hooray!';
     }
 }
 
 class FixturesSiteFake extends Site
 {
     private $valetHomePath;
+
     private $crtCounter = 0;
 
-    public function valetHomePath()
+    public function valetHomePath(): string
     {
         if (! isset($this->valetHomePath)) {
             throw new \RuntimeException(static::class.' needs to be configured using useFixtures or useOutput');
@@ -985,7 +1006,7 @@ class FixturesSiteFake extends Site
         $this->valetHomePath = __DIR__.'/output';
     }
 
-    public function createCa($caExpireInDays)
+    public function createCa(int $caExpireInDays): void
     {
         // noop
         //
@@ -994,7 +1015,7 @@ class FixturesSiteFake extends Site
         // CA for our faked Site.
     }
 
-    public function createCertificate($urlWithTld, $caExpireInDays)
+    public function createCertificate(string $urlWithTld, int $caExpireInDays): void
     {
         // We're not actually going to generate a real certificate
         // here. We are going to do something basic to include
@@ -1061,7 +1082,7 @@ class FixturesSiteFake extends Site
 
 class StubForRemovingLinks extends Site
 {
-    public function sitesPath($additionalPath = null)
+    public function sitesPath(?string $additionalPath = null): string
     {
         return __DIR__.'/output'.($additionalPath ? '/'.$additionalPath : '');
     }
